refactor: merge curve_fields.ts into fields.ts
- Consolidate all field types (BN254, Grumpkin, Secp256k1, Secp256r1) into single fields.ts file
- Keep backward compatibility with Fr and Fq aliases for BN254
- Add both value and type exports for Fr to support type annotations
- Remove redundant curve_fields.ts file
- All field types retain branding and conversion methods from previous commit
